Cleanup imports, move Config to env package and remove redundant classes
