+++ core/src/main/kotlin/org/sourcegrade/jagr/core/extra/Unpack.kt
@@ -28,7 +28,7 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.slf4j.Logger
 import org.sourcegrade.jagr.core.testing.SubmissionInfoImpl
-import org.sourcegrade.jagr.launcher.opt.Config
+import org.sourcegrade.jagr.launcher.env.Config
 import java.io.File
 import java.nio.file.FileSystems
 import kotlin.io.path.ExperimentalPathApi
