--- launcher/src/main/kotlin/org/sourcegrade/jagr/launcher/ipc/IPClient.kt
@@ -1,4 +1,0 @@
-package org.sourcegrade.jagr.launcher.ipc
-
-class IPClient {
-}
