--- launcher/src/main/kotlin/org/sourcegrade/jagr/launcher/ipc/IPHost.kt
@@ -1,4 +1,0 @@
-package org.sourcegrade.jagr.launcher.ipc
-
-class IPHost {
-}
